requirements.txt
numpy==1.26.4
matplotlib==3.9.0
pillow==10.2.0
umap-learn==0.5.6